formkeys.idcount really can be a mediumint (probably smaller actually), and fix its old values from buggy code.  New 'discussions' globj_type. And note where Slashdot has been updated.
tracks internal commit f8aa2358d59dc2f6c6a9fd51b9a2ee12ab11aee0

+++ sql/mysql/slashschema_create.sql
@@ -489,7 +489,7 @@ CREATE TABLE formkeys (
 	formkey varchar(20) DEFAULT '' NOT NULL,
 	formname varchar(32) DEFAULT '' NOT NULL,
 	id varchar(30) DEFAULT '' NOT NULL,
-	idcount int UNSIGNED DEFAULT 0 NOT NULL,
+	idcount mediumint UNSIGNED DEFAULT 0 NOT NULL,
 	uid mediumint UNSIGNED NOT NULL,
 	ipid	char(32) DEFAULT '' NOT NULL,
 	subnetid	char(32) DEFAULT '' NOT NULL,
